+++ apps/osim/tests/test_models.py
@@ -5,14 +5,7 @@ from django.utils import timezone
 
 from apps.osim.models import Check, State, Workflow
 from apps.osim.workflow import WorkflowFramework, WorkflowModel
-from osidb.models import (
-    Flaw,
-    FlawImpact,
-    FlawMeta,
-    FlawResolution,
-    FlawSource,
-    FlawType,
-)
+from osidb.models import Flaw, FlawImpact, FlawMeta, FlawSource, FlawType
 from osidb.tests.factories import AffectFactory, FlawFactory, FlawMetaFactory
 
 pytestmark = pytest.mark.unit
@@ -40,8 +33,6 @@ class CheckDescFactory:
         ("has type", "type", FlawType.VULNERABILITY),
         ("has created_dt", "created_dt", timezone.now()),
         ("has updated_dt", "updated_dt", timezone.now()),
-        ("has state", "state", Flaw.FlawState.NEW),
-        ("has resolution", "resolution", FlawResolution.WONTFIX),
         ("has impact", "impact", FlawImpact.MODERATE),
         ("has title", "title", "random title"),
         ("has description", "description", "random description"),